third_party/blink/web_tests/platform/generic/http/tests/security/img-with-failed-cors-check-fails-to-load-expected.txt

CONSOLE ERROR: Access to image at 'http://localhost:8080/security/resources/red200x100.png' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
ALERT: PASS: The error event was called.
This test passes if the image below does not load. 
